refactor: remove letter file
